Added a dummyTable view using the Eclipse GUI builder.
I had a heck of a time getting to GUI builder to run in 64-bit Eclipse on Windows. To make it work you need to download XULrunner and unzip it to C:\xulrunner. Then go add this to the bottom of your eclipse.ini file (in the eclipse folder): -Dorg.eclipse.swt.browser.XULRunnerPath=C:\xulrunner\xulrunner

After that go to edit the preferences in eclipse: Window->Preferences->General (on left) -> WebBrowser. Select Internet Explorer. (yes really :P)

To edit the GUI, right click on DummyTable.java, OpenWith -> Vaadin editor. There will be a design tab and a source tab on the bottom.

If you want to make a new form, go New->Other->Vaadin Composite.