Fixed the NullPointerException on the login page. Turns out that the singleton I made initially never had a way to change it's contained Navigator object to anything but null. Woops.
Fixed now. You can run the application, click login, and it changes the view to the main screen.

I've been reading about Spring Security integration with Vaadin. Because vaadin is so huge, it's not a trivial thing to bolt on spring security. I don't have enough of an understanding of the Eclipse build system (and Geronimo) to know where to put various config files and classes.